Dutch college data

Longitudinal relational measurements and nodal characteristics of Dutch college students, described in van de Bunt, van Duijn, and Snijders (1999). The time interval between the first four measurements was three weeks, whereas the interval between the last three was six weeks.

Format

A list consisting of a socioarray Y and a matrix X of static nodal attributes. The relational measurements range from -1 to 4, indicating the following:

-1 a troubled or negative relationship

0 don't know

1 neutral relationship

2 friendly

3 friendship

4 best friends

Source

http://moreno.ss.uci.edu/data.html#vdb