AME model fitting routine for replicated relational data

An MCMC routine providing a fit to an additive and multiplicative effects (AME) regression model to replicated relational data of various types.

lame(Y,Xdyad=NULL, Xrow=NULL, Xcol=NULL, rvar = !(family=="rrl")
, cvar = TRUE, dcor = !symmetric, nvar=TRUE,  R = 0, family ="nrm",
intercept=!is.element(family,c("rrl","ord")),
symmetric=FALSE,
odmax=NULL, prior=list(), g=NA,
seed = 6886, nscan = 10000, burn = 500, odens = 25, plot=FALSE, print = FALSE, gof=TRUE,
startVals=NULL, periodicSave=FALSE, outFile=NULL, saveInterval=0.25, model.name=NULL)

Arguments

Y

a T length list of n x n relational matrices, where T corresponds to the number of replicates (over time, for example). See family below for various data types.

Xdyad

a T length list of n x n x pd arrays of covariates

Xrow

a T length list of n x pr matrices of nodal row covariates

Xcol

a T length list of n x pc matrices of nodal column covariates

rvar

logical: fit row random effects (asymmetric case)?

cvar

logical: fit column random effects (asymmetric case)?

dcor

logical: fit a dyadic correlation (asymmetric case)?

nvar

logical: fit nodal random effects (symmetric case)?

R

integer: dimension of the multiplicative effects (can be zero)

family

character: one of "nrm","tob","bin","ord","cbin","frn","rrl" - see the details below

intercept

logical: fit model with an intercept?

symmetric

logical: Is the sociomatrix symmetric by design?

odmax

a scalar integer or vector of length n giving the maximum number of nominations that each node may make - used for "frn" and "cbin" families

prior

a list containing hyperparameters for the prior distributions

g

optional scalar or vector for g-prior

seed

random seed

nscan

number of iterations of the Markov chain (beyond burn-in)

burn

burn in for the Markov chain

odens

output density for the Markov chain

plot

logical: plot results while running?

print

logical: print results while running?

gof

logical: calculate goodness of fit statistics?

startVals

List from previous model run containing parameter starting values for new MCMC

periodicSave

logical: indicating whether to periodically save MCMC results

outFile

character vector indicating name and path in which file should be stored if periodicSave is selected. For example, on an Apple OS outFile="~/Desktop/ameFit.rda".

saveInterval

quantile interval indicating when to save during post burn-in phase.

model.name

optional string for model selection output

Value

BETA

posterior samples of regression coefficients

VC

posterior samples of the variance parameters

APM

posterior mean of additive row effects a

BPM

posterior mean of additive column effects b

U

posterior mean of multiplicative row effects u

V

posterior mean of multiplicative column effects v (asymmetric case)

UVPM

posterior mean of UV

ULUPM

posterior mean of ULU (symmetric case)

L

posterior mean of L (symmetric case)

EZ

estimate of expectation of Z matrix

YPM

posterior mean of Y (for imputing missing values)

GOF

observed (first row) and posterior predictive (remaining rows) values of four goodness-of-fit statistics

startVals

Final parameter values from MCMC, can be used as the input for a future model run.

AIC

Akaike Information Criterion (if model.name provided)

BIC

Bayesian Information Criterion (if model.name provided)

model.name

Name of the model (if provided)

Details

This command provides posterior inference for parameters in AME models of independent replicated relational data, assuming one of seven possible data types/models:

"nrm": A normal AME model.

"tob": A tobit AME model for censored continuous data. Values are censored at zero, appropriate for non-negative continuous relational data.

"bin": A binary probit AME model.

"ord": An ordinal probit AME model. An intercept is not identifiable in this model.

"cbin": An AME model for censored binary data. The value of 'odmax' specifies the maximum number of links each row may have.

"frn": An AME model for fixed rank nomination networks. A higher value of the rank indicates a stronger relationship. The value of 'odmax' specifies the maximum number of links each row may have.

"rrl": An AME model based on the row ranks. This is appropriate if the relationships across rows are not directly comparable in terms of scale. An intercept, row random effects and row regression effects are not estimable for this model.

Author

Peter Hoff, Yanjun He, Shahryar Minhas

Examples

data(YX_bin_list) 
fit<-lame(YX_bin_list$Y,YX_bin_list$X,burn=5,nscan=5,odens=1,family="bin")
#> 
#> Starting burn-in period...
#>   |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  25%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================================| 100%
#> Burn-in period complete...
#>   |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  25%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================================| 100%
# you should run the Markov chain much longer than this