Gibbs sampling of additive row and column effects and regression coefficient

Simulates from the joint full conditional distribution of (a,b,beta)

rbeta_ab_fc(Z, Sab, rho, X, mX, mXt, XX, XXt, Xr, Xc, s2 = 1)

Arguments

Z

n X n (latent) normal relational matrix, with multiplicative effects subtracted out

Sab

row and column covariance

rho

dyadic correlation

X

n x n x p covariate array

mX

design matrix (matricizied version of X)

mXt

dyad-transposed design matrix

XX

regression sums of squares

XXt

crossproduct sums of squares

Xr

row sums for X

Xc

column sums for X

s2

dyadic variance

Value

beta

regression coefficients

a

additive row effects

b

additive column effects

Author

Peter Hoff