Sampson's monastery data

Several dyadic variables measured on 18 members of a monastery.

Format

A socioarray whose dimensions represent nominators, nominatees and relations. Each monk was asked to rank up to three other monks on a variety of positive and negative relations. A rank of three indicates the "highest" ranking for a particular relational variable. The relations like_m2 and like_m1 are evaluations of likeing at one and two timepoints previous to when the other relations were measured.

Source

http://moreno.ss.uci.edu/data.html#sampson